fix: removing IntroViewController tests, moving test assertions to view models
