+++ Sources/Errors/GenericErrorViewModel.swift
@@ -9,24 +9,24 @@ struct GenericErrorViewModel: GDSErrorViewModel, BaseViewModel {
     let primaryButtonViewModel: ButtonViewModel
     let secondaryButtonViewModel: ButtonViewModel? = nil
     let analyticsService: AnalyticsService
-
+    
     let rightBarButtonTitle: GDSLocalisedString? = nil
     let backButtonIsHidden: Bool = true
-
+    
     init(analyticsService: AnalyticsService, action: @escaping () -> Void) {
         self.analyticsService = analyticsService
         self.primaryButtonViewModel = AnalyticsButtonViewModel(titleKey: "Close",
-                                                 analyticsService: analyticsService) {
+                                                               analyticsService: analyticsService) {
             action()
         }
     }
     
     func didAppear() {
         let screen = ScreenView(screen: ErrorAnalyticsScreen.generic,
-                            titleKey: title.stringKey)
+                                titleKey: title.stringKey)
         analyticsService.trackScreen(screen)
     }
-
+    
     func didDismiss() {
         // Here for BaseViewModel compliance
     }
