+++ Tests/UnitTests/Errors/GenericErrorViewModelTests.swift
@@ -1,3 +1,4 @@
+import GDSAnalytics
 @testable import OneLogin
 import XCTest
 
@@ -33,7 +34,23 @@ extension GenericErrorViewModelTests {
 
     func test_buttonAction() throws {
         XCTAssertFalse(didCallButtonAction)
+        XCTAssertEqual(mockAnalyticsService.eventsLogged.count, 0)
         sut.primaryButtonViewModel.action()
         XCTAssertTrue(didCallButtonAction)
+        XCTAssertEqual(mockAnalyticsService.eventsLogged.count, 1)
+        let event = ButtonEvent(textKey: sut.primaryButtonViewModel.title.value)
+        XCTAssertEqual(mockAnalyticsService.eventsLogged, [event.name.name])
+        XCTAssertEqual(mockAnalyticsService.eventsParamsLogged["text"], event.text.lowercased())
+        XCTAssertEqual(mockAnalyticsService.eventsParamsLogged["type"], event.type.rawValue)
+    }
+    
+    func test_didAppear() throws {
+        XCTAssertEqual(mockAnalyticsService.screensVisited.count, 0)
+        sut.didAppear()
+        XCTAssertEqual(mockAnalyticsService.screensVisited.count, 1)
+        let screen = ScreenView(screen: ErrorAnalyticsScreen.generic,
+                                titleKey: "something went wrong")
+        XCTAssertEqual(mockAnalyticsService.screensVisited, [screen.screen.name])
+        XCTAssertEqual(mockAnalyticsService.screenParamsLogged["title"], screen.title)
     }
 }
