+++ Tests/UnitTests/Onboarding/Welcome/OneLoginIntroViewModelTests.swift
@@ -1,3 +1,4 @@
+import GDSAnalytics
 @testable import OneLogin
 import XCTest
 
@@ -34,7 +35,23 @@ extension OneLoginIntroViewModelTests {
     
     func test_buttonAction() throws {
         XCTAssertFalse(didCallButtonAction)
+        XCTAssertEqual(mockAnalyticsService.eventsLogged.count, 0)
         sut.introButtonViewModel.action()
         XCTAssertTrue(didCallButtonAction)
+        XCTAssertEqual(mockAnalyticsService.eventsLogged.count, 1)
+        let event = ButtonEvent(textKey: sut.introButtonViewModel.title.value)
+        XCTAssertEqual(mockAnalyticsService.eventsLogged, [event.name.name])
+        XCTAssertEqual(mockAnalyticsService.eventsParamsLogged["text"], event.text.lowercased())
+        XCTAssertEqual(mockAnalyticsService.eventsParamsLogged["type"], event.type.rawValue)
+    }
+    
+    func test_didAppear() throws {
+        XCTAssertEqual(mockAnalyticsService.screensVisited.count, 0)
+        sut.didAppear()
+        XCTAssertEqual(mockAnalyticsService.screensVisited.count, 1)
+        let screen = ScreenView(screen: IntroAnalyticsScreen.welcomeScreen,
+                                titleKey: "gov.uk one login")
+        XCTAssertEqual(mockAnalyticsService.screensVisited, [screen.screen.name])
+        XCTAssertEqual(mockAnalyticsService.screenParamsLogged["title"], screen.title)
     }
 }
